requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.4.1.post1
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
python-dotenv